Fixed error with wallpaper setting and folders
There was a problem with running the script without setting "-r" resulted in the wallpaper not being set. Turns out this was due to the setWallpaper function not actually being called. Who knew not calling the function to set the wallpaper, wouldn't set the wallpaper!

There was also an issue with the script crashing when the ~/Pictures/Daily didn't exist. This has been solved by the addition of a Python OS call that checks if the directory exists and if not creates it.